Work on the FullThemeSelector, that combines shaders and styles into one.
This includes a port of two new shaders from the Windows Terminal collection
which are under an MIT license.

For the FullThemeSelector, I brought some code from MobileDiffusion (El Pintador) to make the text readable, despite the changing colors or backgrounds
where our text would not show.

This currently contains a simple UI to customize three floating point values
which we could surface to users on their configuration.  Currently these are
global and not per shader, while I work on this.